pyproject.toml
[project]
name = "calabiyau-server-node-switcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dnspython[async]>=2.7.0",
    "fastapi>=0.115.14",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
